Lift state and derive from it in a sub-component
So far, the board was keeping track of the game state.

All of the sudden, we find the need to use that state in a new
component: the log.

So, we lift the game state to the App component, and derive the board
state from that lifted state. We also use that lifted state in a new Log
component that will keep track of all the movements.